requirements.txt
Flask==3.1.0
python-dotenv==1.0.0
upstash-redis==1.1.0
google-generativeai==0.8.3
python-docx==1.1.2
PyPDF2==3.0.1
pillow==10.4.0
pytesseract==0.3.13
tensorflow==2.16.1
Werkzeug>=3.1
gunicorn==22.0.0